pyproject.toml
[project]
name = "insta-dl"
version = "0.1.0"
authors = [
    { name = "pythoninthegrass", email = "4097471+pythoninthegrass@users.noreply.github.com" }
]

requires-python = ">=3.13,<3.14"

dependencies = [
    "browser-cookie3>=0.20.1",
    "instaloader>=4.15",
    "python-decouple>=3.8",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.1",
]
